Extract dehyphenation into a separate library
