+++ src/lib/param-builder.js
@@ -1,5 +1,7 @@
 'use strict';
 
+import {dehyphenate} from './dehyphenate';
+
 export class ParamBuilder {
 
   /**
@@ -10,36 +12,13 @@ export class ParamBuilder {
   }
 
   /**
-  Convert values from hyphenated form to an object tree.
-  */
-  dehyphenate(values) {
-    const dehyphenated = {};
-
-    for (const key in values) {
-      if (values.hasOwnProperty(key)) {
-        const value = values[key];
-        const splitted = key.match(/([^-]+)-+(.*)/);
-
-        if (splitted && this.options.hasOwnProperty(splitted[1])) {
-          dehyphenated[splitted[1]] = dehyphenated[splitted[1]] || {};
-          dehyphenated[splitted[1]][splitted[2]] = value;
-        } else {
-          dehyphenated[key] = value;
-        }
-      }
-    }
-
-    return dehyphenated;
-  }
-
-  /**
   Returns parameters in object form.
   */
   get(values) {
     const params = {};
 
     // Convert "crop-width" to crop.width etc.
-    values = this.dehyphenate(values);
+    values = dehyphenate(values);
 
     for (const key in this.options) {
       if (values.hasOwnProperty(key)) {
